+++ tosixinch/_lxml_html.py
@@ -8,11 +8,14 @@ from tosixinch import system
 
 import lxml.etree as etree
 import lxml.html
-import lxml.html.clean
 import lxml.html.defs
 
+try:
+    import lxml_html_clean as clean  # noqa: F401 (unused)
+except ImportError:
+    import lxml.html.clean as clean  # noqa: F401 (unused)
+
 tostring = lxml.html.tostring
-clean = lxml.html.clean
 link_attrs = lxml.html.defs.link_attrs
 
 logger = logging.getLogger(__name__)